Move crc32c.h to match tensorflow/tensorflow
